Added debug statement

+++ src/read_met_file.c
@@ -302,7 +302,7 @@ void read_subdaily_met_data(char **argv, control *c, met_arrays *ma)
 void read_subdaily_met_data_binary(char **argv, control *c, met_arrays *ma)
 {
     FILE  *fp;
-    double *data = NULL;
+    float *data = NULL;
     char   line[STRING_LENGTH];
     int    i = 0, j = 0;
     int    nvars = 12;
@@ -317,7 +317,7 @@ void read_subdaily_met_data_binary(char **argv, control *c, met_arrays *ma)
 		exit(EXIT_FAILURE);
     }
 
-    if ((data = (double *)calloc(c->nrows*c->ncols, sizeof(double))) == NULL) {
+    if ((data = (float *)calloc(c->nrows*c->ncols, sizeof(float))) == NULL) {
 		fprintf(stderr,"%s: error allocating met data for read\n",argv[0]);
 		exit(EXIT_FAILURE);
 	}
